デコード側対応

+++ cuda_imageprocess.cpp
@@ -7,22 +7,22 @@ extern "C"{
 
     //NV12↔RGBA
     __global__ void nv12_to_rgba_8bit_kernel(uint8_t* rgba, size_t rgba_pitch,const uint8_t* y_plane, size_t y_pitch,const uint8_t* uv_plane, size_t uv_pitch,int width, int height);
-    __global__ void nv12_to_rgba_10bit_kernel(uint8_t* rgba, size_t rgba_pitch,const uint8_t* y_plane, size_t y_pitch,const uint8_t* uv_plane, size_t uv_pitch,int width, int height);
+    __global__ void nv12_to_rgba_10bit_kernel(uint16_t* rgba, size_t rgba_pitch,const uint8_t* y_plane, size_t y_pitch,const uint8_t* uv_plane, size_t uv_pitch,int width, int height);
     __global__ void flip_rgba_to_nv12_kernel(uint8_t* y_plane, size_t y_step,uint8_t* uv_plane,size_t uv_step,const uint8_t* rgba, size_t rgba_step,int width, int height);
 
     //YUV420P→RGBA
     __global__ void yuv420p_to_rgba_8bit_kernel(uint8_t* rgba, size_t rgba_pitch,const uint8_t* y_plane, size_t y_pitch,const uint8_t* u_plane, size_t u_pitch,const uint8_t* v_plane, size_t v_pitch,int width, int height);
-    __global__ void yuv420p_to_rgba_10bit_kernel(uint8_t* rgba, size_t rgba_pitch,const uint8_t* y_plane, size_t y_pitch,const uint8_t* u_plane, size_t u_pitch,const uint8_t* v_plane, size_t v_pitch,int width, int height,int is_be);
+    __global__ void yuv420p_to_rgba_10bit_kernel(uint16_t* rgba, size_t rgba_pitch,const uint8_t* y_plane, size_t y_pitch,const uint8_t* u_plane, size_t u_pitch,const uint8_t* v_plane, size_t v_pitch,int width, int height,int is_be);
 
     //YUV422P→RGBA
     __global__ void yuv422p_to_rgba_8bit_kernel(uint8_t* rgba, size_t rgba_pitch,const uint8_t* y_plane, size_t y_pitch,const uint8_t* u_plane, size_t u_pitch,const uint8_t* v_plane, size_t v_pitch,int width, int height);
-    __global__ void yuv422p_to_rgba_10bit_kernel(uint8_t* rgba, size_t rgba_pitch,const uint8_t* y_plane, size_t y_pitch,const uint8_t* u_plane, size_t u_pitch,const uint8_t* v_plane, size_t v_pitch,int width, int height,int is_be);
+    __global__ void yuv422p_to_rgba_10bit_kernel(uint16_t* rgba, size_t rgba_pitch,const uint8_t* y_plane, size_t y_pitch,const uint8_t* u_plane, size_t u_pitch,const uint8_t* v_plane, size_t v_pitch,int width, int height,int is_be);
 
     //RGB→RGBA
     __global__ void rgb_to_rgba_8bit_kernel(uint8_t* rgba, size_t rgba_pitch,const uint8_t* r_plane, size_t r_pitch,const uint8_t* g_plane, size_t g_pitch,const uint8_t* b_plane, size_t b_pitch,int width, int height);
     __global__ void rgb_to_rgba_8bit_packed_kernel(uint8_t* rgba, size_t rgba_pitch,const uint8_t* src, size_t src_pitch,int width, int height,int mode);
-    __global__ void rgb_to_rgba_10bit_kernel(uint8_t* rgba, size_t rgba_pitch,const uint8_t* r_plane, size_t r_pitch,const uint8_t* g_plane, size_t g_pitch,const uint8_t* b_plane, size_t b_pitch,int width, int height);
-    __global__ void rgb_to_rgba_10bit_packed_kernel(uint8_t* rgba, size_t rgba_pitch,const uint8_t* src, size_t src_pitch,int width, int height,int mode);
+    __global__ void rgb_to_rgba_10bit_kernel(uint16_t* rgba, size_t rgba_pitch,const uint8_t* r_plane, size_t r_pitch,const uint8_t* g_plane, size_t g_pitch,const uint8_t* b_plane, size_t b_pitch,int width, int height);
+    __global__ void rgb_to_rgba_10bit_packed_kernel(uint16_t* rgba, size_t rgba_pitch,const uint8_t* src, size_t src_pitch,int width, int height,int mode);
 
     //特殊(Davinci Resolve YUV8bit)
     __global__ void uyvy422_to_rgba_8bit_kernel(uint8_t* rgba, size_t rgba_pitch,const uint8_t* yuv, size_t yuv_pitch,int width, int height);
@@ -145,7 +145,7 @@ bool CUDA_ImageProcess::NV12_to_RGBA_8bit(uint8_t* d_rgba, size_t pitch_rgba,uin
 }
 
 //NV12→RGBA 10bit
-bool CUDA_ImageProcess::NV12_to_RGBA_10bit(uint8_t* d_rgba, size_t pitch_rgba,uint8_t* d_y, size_t pitch_y,uint8_t* d_uv, size_t pitch_uv,int width,int height, cudaStream_t stream)
+bool CUDA_ImageProcess::NV12_to_RGBA_10bit(uint16_t* d_rgba, size_t pitch_rgba,uint8_t* d_y, size_t pitch_y,uint8_t* d_uv, size_t pitch_uv,int width,int height, cudaStream_t stream)
 {
     void* args[] = {&d_rgba, &pitch_rgba,
                     &d_y, &pitch_y,
@@ -231,7 +231,7 @@ bool CUDA_ImageProcess::yuv420p_to_RGBA_8bit(uint8_t* d_rgba, size_t pitch_rgba,
 }
 
 //YUV420p→RGBA 10bit
-bool CUDA_ImageProcess::yuv420p_to_RGBA_10bit(uint8_t* d_rgba, size_t pitch_rgba,uint8_t* d_y, size_t pitch_y,uint8_t* d_u, size_t pitch_u,uint8_t* d_v, size_t pitch_v,int width, int height,int is_be,cudaStream_t stream){
+bool CUDA_ImageProcess::yuv420p_to_RGBA_10bit(uint16_t* d_rgba, size_t pitch_rgba,uint8_t* d_y, size_t pitch_y,uint8_t* d_u, size_t pitch_u,uint8_t* d_v, size_t pitch_v,int width, int height,int is_be,cudaStream_t stream){
     void* args[] = {&d_rgba, &pitch_rgba,
                     &d_y, &pitch_y,
                     &d_u, &pitch_u,
@@ -291,7 +291,7 @@ bool CUDA_ImageProcess::yuv422p_to_RGBA_8bit(uint8_t* d_rgba, size_t pitch_rgba,
 }
 
 //YUV422p→RGBA 10bit
-bool CUDA_ImageProcess::yuv422p_to_RGBA_10bit(uint8_t* d_rgba, size_t pitch_rgba,uint8_t* d_y, size_t pitch_y,uint8_t* d_u, size_t pitch_u,uint8_t* d_v, size_t pitch_v,int width, int height,int is_be,cudaStream_t stream){
+bool CUDA_ImageProcess::yuv422p_to_RGBA_10bit(uint16_t* d_rgba, size_t pitch_rgba,uint8_t* d_y, size_t pitch_y,uint8_t* d_u, size_t pitch_u,uint8_t* d_v, size_t pitch_v,int width, int height,int is_be,cudaStream_t stream){
     void* args[] = {&d_rgba, &pitch_rgba,
                     &d_y, &pitch_y,
                     &d_u, &pitch_u,
@@ -379,7 +379,7 @@ bool CUDA_ImageProcess::rgb_to_RGBA_8bit_packed(uint8_t* d_rgba, size_t pitch_rg
 }
 
 //RGB→RGBA 10bit planner
-bool CUDA_ImageProcess::rgb_to_RGBA_10bit(uint8_t* d_rgba, size_t pitch_rgba,uint8_t* d_r, size_t pitch_r,uint8_t* d_g, size_t pitch_g,uint8_t* d_b, size_t pitch_b,int width, int height,cudaStream_t stream){
+bool CUDA_ImageProcess::rgb_to_RGBA_10bit(uint16_t* d_rgba, size_t pitch_rgba,uint8_t* d_r, size_t pitch_r,uint8_t* d_g, size_t pitch_g,uint8_t* d_b, size_t pitch_b,int width, int height,cudaStream_t stream){
     void* args[] = {&d_rgba, &pitch_rgba,
                     &d_r, &pitch_r,
                     &d_g, &pitch_g,
